fix deadlock with condition statement
When waiting for condition worker_wakeup_cond to trigger, the worker_wakeup_lock
was grabbed.  If there was a queue for this lock, pthread_cond_broadcast would
*only* unlock the thread holding the lock, and the others would deadlock
(because worker_wakeup_cond was never signalled again).  We fix this by adding a
condition signal before pthread_exit.  That way, every thread in the queue will
get signaled.

Also as part of this fix, make the pthread structs allocated within threadpool,
rather than pointers.  This saves some mallocs.